model: move Model class to own package
