+++ monospace/src/test/java/object/TestObject.java
@@ -2,7 +2,7 @@ package object;
 
 import org.black_matter.monospace.object.GameObject;
 import org.black_matter.monospace.object.components.Renderable;
-import org.black_matter.monospace.render.Model;
+import org.black_matter.monospace.model.Model;
 
 import java.util.HashMap;
